feat: error response handle

+++ src/prisma/subjectPrisma/subject-prisma.service.ts
@@ -24,6 +24,9 @@ export class SubjectPrismaService {
     const classGroupData = await this.prismaService.groupClass.findFirst({
       where: { uid: { contains: data.groupClass.uid } },
     });
+
+    if (!classGroupData) return null;
+
     return this.prismaService.classSubject.create({
       data: {
         title: data.title,
@@ -51,18 +54,36 @@ export class SubjectPrismaService {
   }
 
   async editSubject(where: { id: number }, data: { title: string; description: string; maxScore: number; dueDate: Date }) {
-    return this.prismaService.classSubject.update({
+    let subjectData = await this.prismaService.classSubject.findFirst({ where });
+
+    if (!subjectData) return null;
+
+    subjectData = await this.prismaService.classSubject.update({
       where,
       data: {
         title: data.title,
         description: data.description,
         maxScore: data.maxScore,
         dueDateAt: data.dueDate,
+        status: 'OPEN',
       },
     });
+
+    await this.prismaService.userTask.updateMany({ where: { classSubjectId: where.id }, data: { status: 'PENDING' } });
+    return subjectData;
   }
 
   async updateStatusSubject(where: { id: number }, data: { status: statusSubject }) {
+    const userTasksData = await this.prismaService.userTask.findMany({ where: { classSubjectId: where.id }, include: { fileTask: true } });
+
+    for (const userTask of userTasksData) {
+      if (!userTask.fileTask.length) {
+        await this.prismaService.userTask.update({ where: { id: userTask.id }, data: { status: 'NOT_COLLECTING' } });
+      } else if (userTask.fileTask.length) {
+        await this.prismaService.userTask.update({ where: { id: userTask.id }, data: { status: 'COMPLATE' } });
+      }
+    }
+
     return this.prismaService.classSubject.update({
       where,
       data,
@@ -70,6 +91,10 @@ export class SubjectPrismaService {
   }
 
   async deleteSubject(where: { id: number }) {
+    const subjectData = await this.prismaService.classSubject.findFirst({ where });
+
+    if (!subjectData) return null;
+
     await this.prismaService.fileTask.deleteMany({ where: { userTask: { classSubjectId: where.id } } });
     await this.prismaService.userTask.deleteMany({ where: { classSubjectId: where.id } });
     return this.prismaService.classSubject.delete({ where });
